+++ packages/terra-menu/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed issue where toggled item on nested menus was being reset while navigating back and forth between them.
+
 ## 6.69.0 - (June 12, 2023)
 
 * Changed