+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -8,6 +8,7 @@
   * Added masked cell support.
   * Added hover state for selectable cells.
   * Added blank cell support.
+  * Added ability to allow consumers to specify the sort indicator and error state for columns.
 
 ## 0.1
   * Initial stable release.